requirements.txt
Django
djangorestframework
djangorestframework-hstore
dj-database-url
psycopg2
raven
django-grappelli
gunicorn
django-filter
dj-static
celery
django-celery
redis